requirements.txt
Flask>=2.2.0
Flask-SQLAlchemy>=2.5.0
Flask-Login>=0.6.0
Flask-WTF>=1.0.0
Flask-Admin>=1.6.0
Flask-Bcrypt>=1.0.0
Pillow>=10.0.0
Werkzeug>=2.2.0
Jinja2>=3.1.0
SQLAlchemy>=1.4.0
email-validator
Flask-Mail>=0.9.1
itsdangerous>=2.1.2
packaging